backend/requirements.txt
mcp[cli]
uvicorn
fastapi
crewai
crewai-tools
openai
tavily-python
arxiv
trafilatura
pymupdf
requests
